protocols/tcp: Correct implementation of MSG_NOSIGNAL.
The other send_signal invocation correctly checked NOSIGNAL already.

+++ src/add-ons/kernel/network/protocols/tcp/TCPEndpoint.cpp
@@ -846,7 +846,7 @@ TCPEndpoint::SendData(net_buffer *buffer)
 
 			if (!is_writable(fState) && !is_establishing(fState)) {
 				// we only send signals when called from userland
-				if (gStackModule->is_syscall())
+				if (gStackModule->is_syscall() && (flags & MSG_NOSIGNAL) == 0)
 					send_signal(find_thread(NULL), SIGPIPE);
 				return EPIPE;
 			}
